requirements.txt
alabaster==0.7.12
anaconda-client==1.9.0
anaconda-project==0.10.2
anyio==2.2.0
appdirs==1.4.4
applaunchservices==0.2.1
appnope==0.1.2
appscript==1.1.2
argh==0.26.2
argon2-cffi==20.1.0
arrow==0.13.1
asn1crypto==1.4.0
astroid==2.6.6
astropy==5.0
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.10.0
binaryornot==0.4.4
bitarray==2.3.0
bkcharts==0.2
black==19.10b0
bleach==4.0.0
bokeh==2.4.2
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
certifi==2022.6.15
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.3
cloudpickle==2.0.0
clyent==1.2.2
colorama==0.4.4
conda==4.11.0
conda-content-trust==0+unknown
conda-pack==0.6.0
conda-package-handling==1.7.3
conda-token==0.3.0
contextlib2==0.6.0.post1
cookiecutter==1.7.2
cryptography==36.0.0
cycler==0.11.0
Cython==0.29.25
cytoolz==0.11.0
daal4py==2021.4.0
dask==2021.10.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
diff-match-patch==20200713
distributed==2021.10.0
docutils==0.17.1
entrypoints==0.3
et-xmlfile==1.1.0
fastcache==1.1.0
filelock==3.4.0
flake8==3.9.2
Flask==1.1.2
fonttools==4.25.0
fsspec==2021.10.1
gevent==21.8.0
glob2==0.7
gmpy2==2.0.8
greenlet==1.1.1
gym==0.21.0
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.1
idna==3.3
imagecodecs==2021.8.26
imageio==2.9.0
imagesize==1.3.0
importlib-metadata==4.8.2
inflection==0.5.1
iniconfig==1.1.1
intervaltree==3.1.0
ipykernel==6.4.1
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
itsdangerous==2.0.1
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.3
jinja2-time==0.2.0
joblib==1.1.0
jos3==0.2.6
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyter-server==1.4.1
jupyterlab==3.2.1
jupyterlab-code-formatter==1.5.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.0
keyring==23.4.0
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
libarchive-c==2.9
llvmlite==0.37.0
locket==0.2.1
lxml==4.7.1
MarkupSafe==1.1.1
matplotlib==3.5.0
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mock==4.0.3
more-itertools==8.12.0
mpmath==1.2.1
msgpack==1.0.2
multipledispatch==0.6.0
munkres==1.1.4
mypy-extensions==0.4.3
natsort==8.1.0
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
nltk==3.6.5
nose==1.3.7
notebook==6.4.6
numba==0.54.1
numexpr==2.8.1
numpy==1.20.3
numpydoc==1.1.0
olefile==0.46
openpyxl==3.0.9
packaging==21.3
pandas==1.3.4
pandocfilters==1.4.3
parso==0.8.2
partd==1.2.0
path==16.0.0
pathlib2==2.3.6
pathspec==0.7.0
patsy==0.5.2
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip==21.2.4
pkginfo==1.7.1
pluggy==1.0.0
ply==3.11
poyo==0.5.0
prometheus-client==0.12.0
prompt-toolkit==3.0.20
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.7.0
pycosat==0.6.3
pycparser==2.21
pycurl==7.44.1
pydocstyle==6.1.1
pyerfa==2.0.0
pyflakes==2.3.1
Pygments==2.10.0
pylint==2.9.6
pyls-spyder==0.4.0
pyodbc==4.0.0-unsupported
pyOpenSSL==21.0.0
pyparsing==3.0.4
pyrsistent==0.18.0
PySocks==1.7.1
pytest==6.2.5
pythermalcomfort==1.10.0
python-dateutil==2.8.2
python-lsp-black==1.0.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.4
python-slugify==5.0.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.3.0
QDarkStyle==3.0.2
qstylizer==0.1.10
QtAwesome==1.0.3
qtconsole==5.1.1
QtPy==1.10.0
regex==2021.8.3
requests==2.26.0
rope==0.21.1
Rtree==0.9.7
ruamel-yaml-conda==0.15.100
scikit-image==0.18.3
scikit-learn==1.0.2
scikit-learn-intelex==2021.20211201.94539
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==58.0.4
simplegeneric==0.8.1
singledispatch==3.7.0
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcollections==2.1.0
sortedcontainers==2.4.0
soupsieve==2.3.1
Sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-websupport==1.2.4
spyder==5.1.5
spyder-kernels==2.1.3
SQLAlchemy==1.4.27
statsmodels==0.12.2
sympy==1.9
tables==3.6.1
TBB==0.2
tblib==1.7.0
terminado==0.9.4
testpath==0.5.0
text-unidecode==1.3
textdistance==4.2.1
threadpoolctl==2.2.0
three-merge==0.1.1
tifffile==2021.7.2
tinycss==0.4
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typed-ast==1.4.3
typing-extensions==3.10.0.2
ujson==4.2.0
unicodecsv==0.14.1
Unidecode==1.2.0
urllib3==1.26.7
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wheel==0.37.0
whichcraft==0.6.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wurlitzer==3.0.2
xlrd==2.0.1
XlsxWriter==3.0.2
xlwings==0.24.9
xlwt==1.3.0
yapf==0.31.0
zict==2.0.0
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
